# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDiscourse API
  description: |
    Enterprise-grade API for government document analysis, retrieval, and AI-powered research.
    
    ## Features
    - Document ingestion and processing
    - Semantic search with vector embeddings
    - AI-powered question answering
    - Entity extraction and relationship mapping
    - Task and inference management
    - System diagnostics and monitoring
    
    ## Authentication
    Most endpoints require authentication via API key or JWT token.
    
  version: 1.0.0
  contact:
    name: OpenDiscourse Team
    email: support@opendiscourse.org
    url: https://github.com/cbwinslow/opendiscourse
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.opendiscourse.org
    description: Production server

security:
  - ApiKeyAuth: []
  - BearerAuth: []

paths:
  /v1/search:
    get:
      summary: Search documents
      description: Search across the document corpus using semantic or keyword search
      tags:
        - Search
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            example: "infrastructure spending"
        - name: mode
          in: query
          description: Search mode
          schema:
            type: string
            enum: [semantic, keyword, hybrid]
            default: semantic
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: min_score
          in: query
          description: Minimum relevance score
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0.0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/ingest/documents:
    post:
      summary: Upload documents
      description: Upload and process documents for ingestion into the system
      tags:
        - Ingestion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                metadata:
                  type: string
                  description: JSON metadata for the documents
              required:
                - files
      responses:
        '200':
          description: Documents uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResponse'
        '400':
          description: Invalid file format or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/ingest/tasks:
    post:
      summary: Ingest tasks
      description: Create and store tasks in the system
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskIngestionRequest'
      responses:
        '200':
          description: Tasks ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskIngestionResponse'
        '400':
          description: Invalid task data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/ingest/inferences:
    post:
      summary: Ingest inferences
      description: Create and store AI-generated inferences
      tags:
        - Inferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceIngestionRequest'
      responses:
        '200':
          description: Inferences ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferenceIngestionResponse'

  /v1/chat:
    post:
      summary: AI-powered chat
      description: Send a question and receive AI-generated response with sources
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'

  /diagnostics/linux:
    get:
      summary: Get Linux system diagnostics
      description: Collect comprehensive system diagnostics from Linux systems
      tags:
        - Diagnostics
      responses:
        '200':
          description: System diagnostics collected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsResponse'
        '500':
          description: Diagnostics collection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ingest/diagnostics:
    post:
      summary: Store diagnostics report
      description: Store system diagnostics report in the database
      tags:
        - Diagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosticsIngestionRequest'
      responses:
        '200':
          description: Diagnostics report stored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsIngestionResponse'

  /v1/entities:
    get:
      summary: List entities
      description: Retrieve extracted entities with optional filtering
      tags:
        - Entities
      parameters:
        - name: type
          in: query
          description: Filter by entity type
          schema:
            type: string
            enum: [PERSON, ORGANIZATION, LOCATION, GOVERNMENT_BODY]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of entities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitiesResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
          description: Total number of results
        page:
          type: integer
          description: Current page number
        query:
          type: string
          description: Original search query
        took:
          type: number
          description: Search time in milliseconds

    SearchResult:
      type: object
      properties:
        id:
          type: string
          description: Document ID
        title:
          type: string
          description: Document title
        content:
          type: string
          description: Document content excerpt
        score:
          type: number
          minimum: 0
          maximum: 1
          description: Relevance score
        source:
          type: string
          description: Document source
        entities:
          type: array
          items:
            type: string
          description: Extracted entities
        highlights:
          type: array
          items:
            type: string
          description: Highlighted terms
        metadata:
          type: object
          description: Additional document metadata

    TaskIngestionRequest:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        metadata:
          type: object
          description: Optional metadata
      required:
        - tasks

    Task:
      type: object
      properties:
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        status:
          type: string
          enum: [todo, in-progress, completed, cancelled]
          default: todo
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        assignee:
          type: string
          description: Task assignee
        due_date:
          type: string
          format: date-time
          description: Task due date
        metadata:
          type: object
          description: Additional task metadata
        entity_ids:
          type: array
          items:
            type: string
          description: Related entity IDs
        document_ids:
          type: array
          items:
            type: string
          description: Related document IDs

    TaskIngestionResponse:
      type: object
      properties:
        success:
          type: boolean
        task_ids:
          type: array
          items:
            type: string
        metadata:
          type: object

    InferenceIngestionRequest:
      type: object
      properties:
        inferences:
          type: array
          items:
            $ref: '#/components/schemas/Inference'
        metadata:
          type: object
      required:
        - inferences

    Inference:
      type: object
      properties:
        type:
          type: string
          description: Inference type
        content:
          type: string
          description: Inference content
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score
        source_document_id:
          type: string
          description: Source document ID
        source_task_id:
          type: string
          description: Source task ID
        metadata:
          type: object
        entity_ids:
          type: array
          items:
            type: string
      required:
        - type
        - content

    InferenceIngestionResponse:
      type: object
      properties:
        success:
          type: boolean
        inference_ids:
          type: array
          items:
            type: string
        metadata:
          type: object

    ChatRequest:
      type: object
      properties:
        message:
          type: string
          description: User message/question
        context:
          type: array
          items:
            type: string
          description: Additional context documents
        mode:
          type: string
          enum: [qa, chat, research]
          default: qa
      required:
        - message

    ChatResponse:
      type: object
      properties:
        response:
          type: string
          description: AI-generated response
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        confidence:
          type: number
          description: Response confidence score
        took:
          type: number
          description: Response time in milliseconds

    Source:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        excerpt:
          type: string
        score:
          type: number

    DiagnosticsResponse:
      type: object
      properties:
        hardware:
          $ref: '#/components/schemas/HardwareInfo'
        network:
          type: array
          items:
            $ref: '#/components/schemas/NetworkInterface'
        processes:
          type: array
          items:
            type: object
        kernel:
          type: object
        collection_errors:
          type: array
          items:
            type: string

    HardwareInfo:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        cpu:
          $ref: '#/components/schemas/CPUInfo'
        memory:
          $ref: '#/components/schemas/MemoryInfo'
        disks:
          type: array
          items:
            $ref: '#/components/schemas/DiskInfo'
        raw_smart:
          type: string

    CPUInfo:
      type: object
      properties:
        model:
          type: string
        architecture:
          type: string
        cores:
          type: integer
        sockets:
          type: integer
        threads_per_core:
          type: integer
        max_frequency:
          type: string
        temperatures:
          type: array
          items:
            type: number

    MemoryInfo:
      type: object
      properties:
        total:
          type: string
        available:
          type: string
        swap_total:
          type: string
        swap_used:
          type: string

    DiskInfo:
      type: object
      properties:
        name:
          type: string
        size:
          type: string
        type:
          type: string
        mountpoint:
          type: string

    NetworkInterface:
      type: object
      properties:
        name:
          type: string
        state:
          type: string
        ipv4:
          type: string
        speed:
          type: string

    DiagnosticsIngestionRequest:
      type: object
      properties:
        report:
          type: object
          description: Diagnostics report data
        host:
          type: string
          description: Host identifier
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
      required:
        - report

    DiagnosticsIngestionResponse:
      type: object
      properties:
        report_id:
          type: string
        status:
          type: string

    EntitiesResponse:
      type: object
      properties:
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        total:
          type: integer
        page:
          type: integer

    Entity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    IngestionResponse:
      type: object
      properties:
        success:
          type: boolean
        document_ids:
          type: array
          items:
            type: string
        processed_count:
          type: integer
        failed_count:
          type: integer
        errors:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
      required:
        - error

tags:
  - name: Search
    description: Document search and retrieval
  - name: Ingestion
    description: Document and data ingestion
  - name: Tasks
    description: Task management
  - name: Inferences
    description: AI inference management
  - name: Chat
    description: AI-powered chat and Q&A
  - name: Diagnostics
    description: System diagnostics and monitoring
  - name: Entities
    description: Entity management and extraction